+++ R/theme_who.R
@@ -1,9 +1,17 @@
 #' World Health Organization-inspired ggplot2 theme
 #' 
-#' @rdname theme_who
-#' @export
 #' 
-#' @seealso [bbplot::bbc_style()]
+#' @description This theme is inspired by that used in the annual global Tuberculosis report. 
+#' See the report [here](https://www.who.int/tb/publications/global_report/en/).
+#' @export
+#' @importFrom ggplot2 element_text theme element_blank margin element_line
+#' @seealso bbplot::bbc_style()
+#' @author Maria Bekker-Nielsen Dunbar
+#' @author Sam Abbott
+#' @examples
+#' 
+#' plot_tb_burden_summary(conf = NULL, verbose = FALSE) +
+#' theme_who()
 theme_who <- function(){
   font <- "sans" 
   sz <- 20
@@ -25,13 +33,6 @@ theme_who <- function(){
                                          colour = "black",
                                          hjust = 1),
     
-    # Legend
-    legend.position = "none",
-    legend.text.align = 0,
-    legend.background = ggplot2::element_blank(),
-    legend.title = ggplot2::element_blank(),
-    legend.key = ggplot2::element_blank(),
-    legend.text = ggplot2::element_blank(),
     
     # Axis format
     axis.title = ggplot2::element_text(family = font,
@@ -59,59 +60,4 @@ theme_who <- function(){
     # Remove boxes around facet titles
     strip.text = ggplot2::element_text(size = sz + 4, hjust = 0)
   )
-}
-
-who_pal <- list("main" = c("#000000", "#EE1A24", "#1591D1"),
-                "light" = c("#000000", "#FCBEA7", "#BDD5EF"))
-
-#' World Health Organization-inspired palettes
-#' 
-#' @param palette The colours of choice, current options are main and light
-#' @param reverse 
-#' 
-#' @export
-#' 
-#' @seealso scale_colour_who scale_fill_who
-who_palettes <- function(palette = "main", reverse = FALSE, ...){
-  pal <- who_pal[[palette]]
-  if (reverse) pal <- rev(pal)
-  colorRampPalette(pal, ...)
-}
-
-#' World Health Organization-inspired colours
-#' 
-#' @param palette The colours of choice, current options are main and light
-#' @param reverse 
-#' 
-#' @rdname scale_colour_who
-#' @export
-#' 
-#' @seealso [scale_colour_who() scale_fill_who() who_palettes()]
-scale_colour_who <- function(palette = "main", discrete = TRUE,
-                             reverse = FALSE, ...) {
-  pal <- who_palettes(palette = palette, reverse = reverse)
-  if (isTRUE(discrete)){
-    ggplot2::discrete_scale("colour", paste0("who_", palette), palette = pal, ...)
-  } else {
-    ggplot2::scale_colour_viridis_c(...)
-  }
-}
-
-#' @rdname scale_colour_who
-#' @export
-#' @usage NULL
-
-scale_color_who <- scale_colour_who
-
-#' @rdname scale_colour_who
-#' @export
-
-scale_fill_who <- function(palette = "light", discrete = TRUE, reverse = FALSE, ...) {
-  pal <- who_palettes(palette = palette, reverse = reverse)
-  
-  if (isTRUE(discrete)){
-    ggplot2::discrete_scale("fill", paste0("who_", palette), palette = pal, ...)
-  } else {
-    ggplot2::scale_fill_viridis_d(...)
-  }
 }