+++ R/map_tb_burden.R
@@ -63,7 +63,7 @@ map_tb_burden <- function(df = NULL, dict = NULL,
                            viridis_palette = "viridis",
                            viridis_direction = -1,
                            viridis_end = 0.9,
-                           verbose = TRUE, ...) {
+                           verbose = FALSE, ...) {
   
   long <- NULL; lat <- NULL; Year <- NULL; 
   if (!is.null(facet) && facet %in% "year") {