+++ R/search_data_dict.R
@@ -42,7 +42,7 @@ search_data_dict <- function(var = NULL, def = NULL, dataset = NULL,
                              dict = NULL, 
                              download_data = TRUE, save = TRUE, 
                              dict_save_name = NULL,
-                             verbose = TRUE, ...) {
+                             verbose = FALSE, ...) {
   if (is.null(var) && is.null(def) && is.null(dataset)) {
     stop("At least one variable name, definition fragment, or dataset should be supplied")
   }